19- add reservation handling and error messaging; improve chambre deletion logic and enhance reservation factory

+++ app/Http/Controllers/ChambreController.php
@@ -146,6 +146,10 @@ class ChambreController extends Controller
      */
     public function destroy(Chambre $chambre)
     {
+        if($chambre->reservations->count()){
+            return redirect()->route('chambres.index')->with('error', 'Opération interdite : chambre déjà un objet de réservation.');
+        }
+
         $chambre->delete() ;
         return redirect()->route('chambres.index')->with('success', 'Chambre supprimer avec success!');
     }
